add playlist subcommand

+++ src/index.ts
@@ -7,14 +7,13 @@ import {
   readFileSync,
   writeFileSync,
   mkdirSync,
-  lchown,
-  rmdirSync,
   rm,
 } from "node:fs";
 import path from "node:path";
-import { navigateArray } from "./arrayNavigation";
-import { writeFile } from "node:fs/promises";
+import { navigateArray, type ArrayNavigationResult } from "./arrayNavigation";
+import { exists, writeFile } from "node:fs/promises";
 import * as readline from "readline";
+import FuzzySearch from "fuzzy-search";
 
 const configPath = path.join(import.meta.dirname, "../config.yml");
 
@@ -32,6 +31,61 @@ type PingResponse = {
     serverVersion: string;
     openSubsonic: boolean;
   };
+};
+
+// TODO please fix this
+
+type Song = {
+  id: string;
+  parent: string;
+  isDir: boolean;
+  title: string;
+  album: string;
+  artist: string;
+  track: number;
+  year: number;
+  coverArt: string;
+  size: number;
+  contentType: string;
+  suffix: string;
+  duration: number;
+  bitRate: number;
+  path: string;
+  playCount: number;
+  discNumber: number;
+  created: string;
+  albumId: string;
+  artistId: string;
+  type: string;
+  isVideo: boolean;
+  played: string;
+  bpm: number;
+  comment: string;
+  sortName: string;
+  mediaType: string;
+  musicBrainzId: string;
+  genres: Array<any>;
+  replayGain: {
+    trackPeak: number;
+    albumPeak: number;
+  };
+  channelCount: number;
+  samplingRate: number;
+  bitDepth: number;
+  moods: Array<any>;
+  artists: Array<{
+    id: string;
+    name: string;
+  }>;
+  displayArtist: string;
+  albumArtists: Array<{
+    id: string;
+    name: string;
+  }>;
+  displayAlbumArtist: string;
+  contributors: Array<any>;
+  displayComposer: string;
+  explicitStatus: string;
 };
 
 type SearchResponse = {
@@ -147,6 +201,57 @@ type SearchResponse = {
   };
 };
 
+type PlaylistsResponse = {
+  "subsonic-response": {
+    status: string;
+    version: string;
+    type: string;
+    serverVersion: string;
+    openSubsonic: boolean;
+    playlists: {
+      playlist: Array<{
+        id: string;
+        name: string;
+        songCount: number;
+        duration: number;
+        public: boolean;
+        owner: string;
+        created: string;
+        changed: string;
+        coverArt: string;
+        comment?: string;
+      }>;
+    };
+  };
+};
+
+type GetPlaylistResponse = {
+  "subsonic-response": {
+    status: string;
+    version: string;
+    type: string;
+    serverVersion: string;
+    openSubsonic: boolean;
+    playlist: {
+      id: string;
+      name: string;
+      comment: string;
+      songCount: number;
+      duration: number;
+      public: boolean;
+      owner: string;
+      created: string;
+      changed: string;
+      coverArt: string;
+      entry: Song[];
+    };
+  };
+};
+
+function sanitizeFilename(fileName: string): string {
+  return fileName.replace(/[<>:"/\\|?*\x00-\x1F]/g, "").replace(/\s+/g, "_");
+}
+
 function isValidURL(url: string) {
   // the URL class constructor errors if the passed in string is not a valid URL
   try {
@@ -173,7 +278,7 @@ function createURL(endpoint: string): URL | undefined {
       console.log("Error - credentials not set");
       return;
     }
-    const url = new URL(endpoint, yamlData.url);
+    const url = new URL(`/rest/${endpoint}`, yamlData.url);
     const salt = randomBytes(6).toString("hex");
     const hash = md5(yamlData.password + salt);
 
@@ -182,30 +287,60 @@ function createURL(endpoint: string): URL | undefined {
     url.searchParams.set("s", salt);
     url.searchParams.set("v", "1.16.1");
     url.searchParams.set("c", "scli");
+    url.searchParams.set("f", "json");
 
     return url;
   }
 }
 
-async function playSong(songId: string, songName: string) {
-  const streamURL = createURL("/rest/stream") as URL;
+async function playSong(
+  songId: string,
+  songName: string,
+  playlist: boolean = false
+) {
+  const streamURL = createURL("stream") as URL;
   streamURL.searchParams.set("format", "mp3");
   streamURL.searchParams.set("id", songId);
 
-  mkdirSync(path.join(import.meta.dirname, "../cache"), { recursive: true });
-  await fetch(streamURL)
-    .then(async (data) => await data.arrayBuffer())
-    .then(
-      async (buffer) =>
-        await writeFile(
-          path.join(
-            import.meta.dirname,
-            `../cache/${songName.replaceAll(" ", "_")}.mp3`
-          ),
-          Buffer.from(buffer)
-        )
-    );
-  console.log(`Playing ${songName}... (ESC/Ctrl+C to exit)`);
+  readline.emitKeypressEvents(process.stdin);
+  if (process.stdin.isTTY) process.stdin.setRawMode(true);
+
+  let exit = false;
+
+  const keypressHandler = (_str: string, _key: readline.Key) => {
+    if (playlist) exit = true;
+  };
+
+  process.stdin.on("keypress", keypressHandler);
+
+  if (exit) return;
+
+  if (
+    !existsSync(
+      path.join(
+        import.meta.dirname,
+        `../cache/${sanitizeFilename(songName.replaceAll(" ", "_"))}.mp3`
+      )
+    )
+  ) {
+    mkdirSync(path.join(import.meta.dirname, "../cache"), { recursive: true });
+    await fetch(streamURL)
+      .then(async (data) => await data.arrayBuffer())
+      .then(
+        async (buffer) =>
+          await writeFile(
+            path.join(
+              import.meta.dirname,
+              `../cache/${sanitizeFilename(songName.replaceAll(" ", "_"))}.mp3`
+            ),
+            Buffer.from(buffer)
+          )
+      );
+  }
+  console.log(`. playing ${!playlist ? "(ESC/Ctrl+C to exit)" : ""}`);
+
+  process.stdin.removeListener("keypress", keypressHandler);
+
   const proc = Bun.spawn([
     "ffplay",
     "-nodisp",
@@ -214,20 +349,19 @@ async function playSong(songId: string, songName: string) {
     "quiet",
     path.join(
       import.meta.dirname,
-      `../cache/${songName.replaceAll(" ", "_")}.mp3`
+      `../cache/${sanitizeFilename(songName.replaceAll(" ", "_"))}.mp3`
     ),
   ]);
 
-  readline.emitKeypressEvents(process.stdin);
-  if (process.stdin.isTTY) process.stdin.setRawMode(true);
+  const killSongHandler = (_str: string, key: readline.Key) => {
+    if (key.name === "escape" || (key.ctrl && key.name === "c")) process.exit();
+    if (playlist) proc.kill();
+  };
 
-  process.stdin.on("keypress", (_str: string, key: readline.Key) => {
-    if (key.name === "escape" || (key.ctrl && key.name === "c")) {
-      proc.kill();
-    }
-  });
+  process.stdin.on("keypress", killSongHandler);
 
   await proc.exited;
+  process.stdin.removeListener("keypress", killSongHandler);
 }
 
 const app = new Command();
@@ -324,8 +458,7 @@ app
         return;
       }
 
-      const pingURL = createURL("/rest/ping") as URL;
-      pingURL.searchParams.set("f", "json");
+      const pingURL = createURL("ping") as URL;
 
       try {
         const ping = (await fetch(pingURL).then(
@@ -345,7 +478,7 @@ app
 
 app
   .command("play")
-  .description("Play album or playlist")
+  .description("Play song via ffplay")
   .argument("<category>", "Type of music to play (album, playlist)")
   .argument("<string>", "Search string")
   .action(async (category: string, string: string) => {
@@ -370,8 +503,7 @@ app
       switch (category) {
         case "s":
         case "song":
-          const searchURL = createURL("/rest/search2") as URL;
-          searchURL.searchParams.set("f", "json");
+          const searchURL = createURL("search2") as URL;
 
           process.stdout.write("Establishing connection to server..");
 
@@ -411,8 +543,72 @@ app
             selectedSong?.id as string,
             selectedSong?.title as string
           );
-          process.exit();
+          break;
+
+        case "p":
+        case "playlist":
+          const playlistURL = createURL("getPlaylists") as URL;
+          try {
+            const playlistsResponse = (await fetch(playlistURL).then(
+              async (res) => res.json()
+            )) as PlaylistsResponse;
+
+            const playlists =
+              playlistsResponse["subsonic-response"].playlists.playlist;
+            const playlistNames = playlists.map((playlist) => playlist.name);
+
+            const playlistSearchResult = new FuzzySearch(playlists, [
+              "name",
+            ]).search(string);
+            let selected: ArrayNavigationResult;
+
+            if (playlistSearchResult.length === 0) {
+              console.log("No search results found");
+              break;
+            } else if (playlistSearchResult.length === 1) {
+              selected = {
+                selectedItem: playlistSearchResult[0]?.name,
+                selectedIndex: playlists.indexOf(playlistSearchResult[0]!),
+                cancelled: false,
+              };
+            } else {
+              selected = await navigateArray(playlistNames);
+            }
+
+            const playlistId = playlists[selected.selectedIndex as number]?.id;
+            if (!playlistId) {
+              console.log("An error occurred while fetching playlist data");
+              break;
+            }
+
+            const playlistListingURL = createURL("getPlaylist") as URL;
+            playlistListingURL.searchParams.set("id", playlistId);
+
+            // oh god not another nested trycatch
+            try {
+              const playlistInfo = (await fetch(playlistListingURL).then(
+                async (res) => await res.json()
+              )) as GetPlaylistResponse;
+
+              console.log("(Press any key for next song, ESC/Ctrl+C to exit)");
+
+              for (const song of playlistInfo["subsonic-response"].playlist
+                .entry) {
+                process.stdout.write(
+                  `Downloading song ${song.title} - ${song.artist}..`
+                );
+                await playSong(song.id, song.title, true);
+              }
+            } catch (e) {
+              console.log(e);
+              console.log("An error occurred while fetching playlist data");
+            }
+          } catch {
+            console.log("An error occured while connecting to the server");
+          }
+          break;
       }
+      process.exit();
     }
   });
